+++ Block/CategoryMapping.php
@@ -1,6 +1,6 @@
 <?php
 /**
- * Copyright (c) 2019, Nosto Solutions Ltd
+ * Copyright (c) 2020, Nosto Solutions Ltd
  * All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without modification,
@@ -29,7 +29,7 @@
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  *
  * @author Nosto Solutions Ltd <contact@nosto.com>
- * @copyright 2019 Nosto Solutions Ltd
+ * @copyright 2020 Nosto Solutions Ltd
  * @license http://opensource.org/licenses/BSD-3-Clause BSD 3-Clause
  *
  */
@@ -44,8 +44,9 @@ use Magento\Framework\View\Element\Template;
 use Magento\Framework\View\Element\Template\Context;
 use Magento\Store\Model\Store;
 use Magento\Store\Model\StoreManagerInterface;
-use Nosto\Tagging\Logger\Logger as NostoLogger;
+use Nosto\Cmp\Logger\LoggerInterface;
 use Nosto\Tagging\Model\Service\Product\Category\DefaultCategoryService as CategoryBuilder;
+use Nosto\Cmp\Helper\Data as NostoHelperData;
 
 class CategoryMapping extends Template
 {
@@ -58,23 +59,34 @@ class CategoryMapping extends Template
     /** @var CategoryBuilder */
     private $categoryBuilder;
 
-    /** @var NostoLogger */
+    /** @var NostoHelperData */
+    private $nostoHelperData;
+
+    /** @var LoggerInterface */
     private $logger;
 
     /**
      * CategoryMapping constructor.
+     * @param StoreManagerInterface $storeManager
+     * @param CollectionFactory $collectionFactory
+     * @param CategoryBuilder $categoryBuilder
+     * @param NostoHelperData $nostoHelperData
+     * @param Context $context
+     * @param LoggerInterface $logger
      */
     public function __construct(
         StoreManagerInterface $storeManager,
         CollectionFactory $collectionFactory,
         CategoryBuilder $categoryBuilder,
+        NostoHelperData $nostoHelperData,
         Context $context,
-        NostoLogger $logger
+        LoggerInterface $logger
     ) {
         parent::__construct($context);
         $this->storeManager = $storeManager;
         $this->collectionFactory = $collectionFactory;
         $this->categoryBuilder = $categoryBuilder;
+        $this->nostoHelperData = $nostoHelperData;
         $this->logger = $logger;
     }
 
@@ -113,9 +125,12 @@ class CategoryMapping extends Template
 
             $categories = $this->collectionFactory->create()
                 ->addAttributeToSelect('*')
-                ->addAttributeToFilter('include_in_menu', ['eq' => 1])
                 ->addIsActiveFilter()
                 ->setStore($store);
+
+            if (!$this->nostoHelperData->isAllCategoriesMapEnabled($store)) {
+                $categories->addAttributeToFilter('include_in_menu', ['eq' => 1]);
+            }
 
             /** @var Category $category $item */
             foreach ($categories->getItems() as $category) {