+++ Model/Filter/FilterBuilder.php
@@ -1,6 +1,6 @@
 <?php
 /**
- * Copyright (c) 2019, Nosto Solutions Ltd
+ * Copyright (c) 2020, Nosto Solutions Ltd
  * All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without modification,
@@ -29,7 +29,7 @@
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  *
  * @author Nosto Solutions Ltd <contact@nosto.com>
- * @copyright 2019 Nosto Solutions Ltd
+ * @copyright 2020 Nosto Solutions Ltd
  * @license http://opensource.org/licenses/BSD-3-Clause BSD 3-Clause
  *
  */
@@ -44,7 +44,7 @@ use Nosto\NostoException;
 use Nosto\Operation\Recommendation\IncludeFilters;
 use Nosto\Operation\Recommendation\ExcludeFilters;
 use Nosto\Tagging\Helper\Data as NostoHelperData;
-use Nosto\Tagging\Logger\Logger as NostoLogger;
+use Nosto\Cmp\Logger\LoggerInterface;
 
 class FilterBuilder
 {
@@ -60,7 +60,7 @@ class FilterBuilder
     /** @var string */
     private $brand;
 
-    /** @var NostoLogger */
+    /** @var LoggerInterface */
     private $logger;
 
     /**
@@ -68,13 +68,13 @@ class FilterBuilder
      * @param IncludeFilters $includeFilters
      * @param ExcludeFilters $excludeFilters
      * @param NostoHelperData $nostoHelperData
-     * @param NostoLogger $logger
+     * @param LoggerInterface $logger
      */
     public function __construct(
         IncludeFilters $includeFilters,
         ExcludeFilters $excludeFilters,
         NostoHelperData $nostoHelperData,
-        NostoLogger $logger
+        LoggerInterface $logger
     ) {
         $this->includeFilters = $includeFilters;
         $this->excludeFilters = $excludeFilters;
@@ -142,16 +142,19 @@ class FilterBuilder
                 $value = (bool) $item->getData('value');
                 break;
             default:
-                $this->logger->debug(sprintf(
-                    'Cannot build include filter for "%s" frontend input type',
-                    $frontendInput
-                ));
+                $this->logger->debugCmp(
+                    sprintf(
+                        'Cannot build include filter for "%s" frontend input type',
+                        $frontendInput
+                    ),
+                    $this
+                );
                 return;
         }
         try {
             $this->mapValueToFilter($filterName, $value);
         } catch (NostoException $e) {
-            $this->logger->info($e->getMessage());
+            $this->logger->exception($e);
         }
     }