+++ Model/Service/Recommendation/Category.php
@@ -1,6 +1,6 @@
 <?php
 /**
- * Copyright (c) 2019, Nosto Solutions Ltd
+ * Copyright (c) 2020, Nosto Solutions Ltd
  * All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without modification,
@@ -29,7 +29,7 @@
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  *
  * @author Nosto Solutions Ltd <contact@nosto.com>
- * @copyright 2019 Nosto Solutions Ltd
+ * @copyright 2020 Nosto Solutions Ltd
  * @license http://opensource.org/licenses/BSD-3-Clause BSD 3-Clause
  *
  */
@@ -38,6 +38,8 @@ namespace Nosto\Cmp\Model\Service\Recommendation;
 
 use Nosto\Model\Signup\Account as NostoAccount;
 use Nosto\Operation\AbstractGraphQLOperation;
+use Nosto\Request\Http\Exception\AbstractHttpException;
+use Nosto\Request\Http\Exception\HttpResponseException;
 use Nosto\Service\FeatureAccess;
 use Nosto\Operation\Recommendation\CategoryMerchandising;
 use Magento\Framework\Stdlib\CookieManagerInterface;
@@ -69,8 +71,11 @@ class Category
      * @param $category
      * @param int $pageNumber
      * @param int $limit
+     * @param bool $previewMode
      * @return CategoryMerchandisingResult
      * @throws NostoException
+     * @throws AbstractHttpException
+     * @throws HttpResponseException
      */
     public function getPersonalisationResult(
         NostoAccount $nostoAccount,
@@ -78,14 +83,13 @@ class Category
         $nostoCustomerId,
         $category,
         $pageNumber,
-        $limit = self::MAX_PRODUCT_AMOUNT
+        $limit,
+        $previewMode = false
     ) {
-        $limit = self::MAX_PRODUCT_AMOUNT < $limit ? self::MAX_PRODUCT_AMOUNT : $limit;
         $featureAccess = new FeatureAccess($nostoAccount);
         if (!$featureAccess->canUseGraphql()) {
             throw new NostoException('Missing Nosto API_APPS token');
         }
-        $previewMode = (bool)$this->cookieManager->getCookie(self::NOSTO_PREVIEW_COOKIE);
         $categoryMerchandising = new CategoryMerchandising(
             $nostoAccount,
             $nostoCustomerId,
@@ -98,7 +102,6 @@ class Category
             $previewMode,
             $limit
         );
-
         return $categoryMerchandising->execute();
     }
 }