+++ Plugin/Catalog/Block/AbstractBlock.php
@@ -1,6 +1,6 @@
 <?php
 /**
- * Copyright (c) 2019, Nosto Solutions Ltd
+ * Copyright (c) 2020, Nosto Solutions Ltd
  * All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without modification,
@@ -29,7 +29,7 @@
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  *
  * @author Nosto Solutions Ltd <contact@nosto.com>
- * @copyright 2019 Nosto Solutions Ltd
+ * @copyright 2020 Nosto Solutions Ltd
  * @license http://opensource.org/licenses/BSD-3-Clause BSD 3-Clause
  *
  */
@@ -38,46 +38,44 @@ namespace Nosto\Cmp\Plugin\Catalog\Block;
 
 use Magento\Backend\Block\Template\Context;
 use Magento\Catalog\Block\Product\ProductList\Toolbar as MagentoToolbar;
-use Magento\Framework\Exception\NoSuchEntityException;
 use Magento\Framework\View\Element\Template;
-use Magento\Theme\Block\Html\Pager as MagentoPager;
+use Magento\Store\Api\Data\StoreInterface;
 use Magento\Store\Model\StoreManagerInterface;
+use Magento\Theme\Block\Html\Pager as MagentoPager;
 use Nosto\Cmp\Helper\CategorySorting as NostoHelperSorting;
 use Nosto\Cmp\Helper\Data as NostoCmpHelperData;
+use Nosto\Cmp\Model\Service\Recommendation\StateAwareCategoryService;
+use Nosto\Cmp\Model\Service\Recommendation\StateAwareCategoryServiceInterface;
 use Nosto\Tagging\Helper\Account as NostoHelperAccount;
-use Nosto\Tagging\Logger\Logger as NostoLogger;
+use Nosto\Cmp\Logger\LoggerInterface;
 
 abstract class AbstractBlock extends Template
 {
     /** @var int */
-    public static $totalProducts;
-
-    /** @var int */
-    public static $limit;
-
-    /** @var int */
     private $lastPageNumber;
 
     /** @var ParameterResolverInterface */
-    public $paramResolver;
+    private $paramResolver;
 
     /**  @var StoreManagerInterface */
-    public $storeManager;
+    private $storeManager;
 
     /** @var NostoCmpHelperData */
-    public $nostoCmpHelperData;
+    private $nostoCmpHelperData;
 
     /** @var NostoHelperAccount */
-    public $nostoHelperAccount;
+    private $nostoHelperAccount;
 
-    /** @var NostoLogger */
-    public $logger;
+    /** @var LoggerInterface */
+    private $logger;
 
     /** @var string */
     public static $currentOrder;
 
-    /** @var bool */
-    public static $catalogTakeover;
+    /**
+     * @var StateAwareCategoryService
+     */
+    private $categoryService;
 
     /**
      * AbstractBlock constructor.
@@ -85,15 +83,18 @@ abstract class AbstractBlock extends Template
      * @param ParameterResolverInterface $parameterResolver
      * @param NostoCmpHelperData $nostoCmpHelperData
      * @param NostoHelperAccount $nostoHelperAccount
-     * @param NostoLogger $logger
+     * @param StateAwareCategoryServiceInterface $categoryService
+     * @param LoggerInterface $logger
      */
     public function __construct(
         Context $context,
         ParameterResolverInterface $parameterResolver,
         NostoCmpHelperData $nostoCmpHelperData,
         NostoHelperAccount $nostoHelperAccount,
-        NostoLogger $logger
+        StateAwareCategoryServiceInterface $categoryService,
+        LoggerInterface $logger
     ) {
+        $this->categoryService = $categoryService;
         $this->paramResolver = $parameterResolver;
         $this->nostoCmpHelperData = $nostoCmpHelperData;
         $this->nostoHelperAccount = $nostoHelperAccount;
@@ -106,16 +107,11 @@ abstract class AbstractBlock extends Template
      * Checks if current sorting order is Nosto's `Personalized for you`
      * and category sorting is enabled
      *
+     * @param StoreInterface $store
      * @return bool
      */
-    public function isCmpCurrentSortOrder()
+    public function isCmpCurrentSortOrder(StoreInterface $store)
     {
-        try {
-            $store = $this->storeManager->getStore();
-        } catch (NoSuchEntityException $e) {
-            $this->logger->info('Cannot get store');
-            return false;
-        }
         $currentOrder = $this->getCurrentOrder();
         if ($currentOrder === null) {
             return false;
@@ -138,23 +134,13 @@ abstract class AbstractBlock extends Template
      */
     public function isCmpTakingOverCatalog()
     {
-        if (self::$catalogTakeover === null) {
-            self::$catalogTakeover = $this->isCmpResult();
+        $categoryMerchandisingResult = $this->getCategoryService()->getLastResult();
+        if ($categoryMerchandisingResult !== null
+            && !empty($categoryMerchandisingResult->getTrackingCode())
+        ) {
+            return true;
         }
-        return self::$catalogTakeover;
-    }
-
-    /**
-     * @return bool
-     */
-    private function isCmpResult()
-    {
-        if (!$this->isCmpCurrentSortOrder() ||
-            (self::$totalProducts === 0 || self::$totalProducts === null)
-           ) {
-            return false;
-        }
-        return true;
+        return false;
     }
 
     /**
@@ -169,19 +155,14 @@ abstract class AbstractBlock extends Template
     }
 
     /**
-     * @param int $totalProducts
-     */
-    public function setTotalProducts($totalProducts)
-    {
-        self::$totalProducts = $totalProducts;
-    }
-
-    /**
-     * @return int
+     * @return int|null
      */
     public function getTotalProducts()
     {
-        return self::$totalProducts;
+        if ($this->getCategoryService()->getLastResult() !== null) {
+            return $this->getCategoryService()->getLastResult()->getTotalPrimaryCount();
+        }
+        return null;
     }
 
     /**
@@ -253,24 +234,16 @@ abstract class AbstractBlock extends Template
         if ($this->lastPageNumber !== null) {
             return $this->lastPageNumber;
         }
-        $this->lastPageNumber = (int) ceil(self::$totalProducts/self::$limit);
+        $this->lastPageNumber = (int) ceil($this->getTotalProducts() / $this->getLimit());
         return $this->lastPageNumber;
     }
 
     /**
      * @return int
      */
-    public function getLimit()
+    private function getLimit()
     {
-        return self::$limit;
-    }
-
-    /**
-     * @param int $limit
-     */
-    public function setLimit(int $limit)
-    {
-        self::$limit = $limit;
+        return $this->getCategoryService()->getLastUsedLimit();
     }
 
     /**
@@ -279,5 +252,29 @@ abstract class AbstractBlock extends Template
     public function getCurrentPageNumber()
     {
         return $this->paramResolver->getCurrentPage();
+    }
+
+    /**
+     * @return StateAwareCategoryService
+     */
+    public function getCategoryService(): StateAwareCategoryService
+    {
+        return $this->categoryService;
+    }
+
+    /**
+     * @return StoreManagerInterface
+     */
+    public function getStoreManager(): StoreManagerInterface
+    {
+        return $this->storeManager;
+    }
+
+    /**
+     * @return LoggerInterface
+     */
+    public function getLogger(): LoggerInterface
+    {
+        return $this->logger;
     }
 }