+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 All notable changes to this project will be documented in this file. This project adheres to Semantic Versioning.
 
+### 3.0.0
+* Add support for using Elasticsearch as a catalog search engine
+* Introduce possibility to use all categories (not only the ones in navigation) with Nosto's category merchandising
+
 ### 2.0.3
 * Fix the default max product limit configuration and set the default value to be 250 which is the current max products limit in Nosto